Apport des dernieres modifications et, création du fichier USER dans le dossier models
