+++ back/controllers/user.js
@@ -1,5 +1,5 @@
 // Imports fichier et, dépendances
-const {User} = require('../mongo');
+const {User} = require('../models/user');
 const bcrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
