+++ back/mongo.js
@@ -1,7 +1,5 @@
 // Import pour la connexion à la base de données MongoDB
 const mongoose = require('mongoose');
-// Import du package pour la validation d'identifiants unique
-const uniqueValidator = require('mongoose-unique-validator')
 
 // Utilisation fichier .env pour l'uri vers la base de données
 const password = process.env.DB_PASSWORD;
@@ -15,13 +13,4 @@ mongoose
 .then(() => console.log("Connecté à Mongo"))
 .catch ((err) => console.log("Echec de la connexion à Mongo", err));
 
-// Construction d'un schéma typique d'un utilisateur (Informations recueillis), adresse email unique
-const userSchema = new mongoose.Schema({
-    email: {type: String, require: true, unique: true},
-    password: {type: String, required: true},
-});
-userSchema.plugin(uniqueValidator);
-
-const User = mongoose.model("User", userSchema);
-
-module.exports = {mongoose, User};
+module.exports = {mongoose};
